Implemented JQuery for interactive features
Jquery will be used sparing throughout development, as this assigment
will be focused on how I deal with issues, not just using a libary to do
everything for me.

JQuery will only be used to add a nice layer of polish to the
application, or if it features a function that's made better than I can
possably make currently.

I'm not too happy about the number of things that need to be imported
currently, but we'll see what happens!