Tweets fully work. Need only Login and Display
Clicking on get tweets will get the server to communicate with the
twitter API to get my tweets and then will throw them in the console.

All I need to do now is have dashboard users login and have the tweets
display on the approprate box.

API Security is a bit crazy however! :worried: