File management functionality started
I want the dashboard to have a file storage aspect, to allow people to
upload and access files on the go, an element I feel could be extremely
useful for every family.

Currently working on uploading, and then I'd ideally like a view much
like Google Drive to see and select each file as an actual file in the
file management window. but we'll see how we do, as I'm a little slow...

Progress!