Boxes can be swapped by clicking the circle in the top-right and then clicking another box.
Dealing with the pairent.insertBefore is a bit difficult, as it means
that It'll always place the element before the selected element, whereas
I expect the user wants the swapping to just swap elements properly.

I'll have to look around for solutions, but the hard work for this
feature has been done for the most part.